+++ demos/demo-stacked-bar.js
@@ -4,7 +4,7 @@ const d3Selection = require('d3-selection');
 
 const PubSub = require('pubsub-js');
 
-const colors = require('./../src/charts/helpers/colors');
+const colors = require('./../src/helpers/color');
 
 const stackedBarChart = require('./../src/charts/stacked-bar');
 const tooltip = require('./../src/charts/tooltip');
