+++ src/charts/bar.js
@@ -13,7 +13,7 @@ define(function(require) {
 
     const textHelper = require('./helpers/text');
     const {exportChart} = require('./helpers/exportChart');
-    const colorHelper = require('./helpers/colors');
+    const colorHelper = require('../helpers/color');
     const {bar} = require('./helpers/loadingStates');
 
     const PERCENTAGE_FORMAT = '%';
