+++ src/charts/grouped-bar.js
@@ -14,9 +14,9 @@ define(function (require) {
     const d3Transition = require('d3-transition');
 
     const { exportChart } = require('./helpers/exportChart');
-    const colorHelper = require('./helpers/colors');
+    const colorHelper = require('../helpers/color');
     const {bar} = require('./helpers/loadingStates');
-    
+
     const NUMBER_FORMAT = ',f';
     const uniq = (arrArg) => arrArg.filter((elem, pos, arr) => arr.indexOf(elem) == pos);
