+++ src/charts/stacked-bar.js
@@ -15,7 +15,7 @@ define(function(require){
     const d3Transition = require('d3-transition');
 
     const {exportChart} = require('./helpers/exportChart');
-    const colorHelper = require('./helpers/colors');
+    const colorHelper = require('../helpers/color');
     const {bar} = require('./helpers/loadingStates');
 
     const PERCENTAGE_FORMAT = '%';
